dom/imptests/html/dom/nodes/test_Document-getElementsByTagNameNS.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.getElementsByTagNameNS</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-getelementsbytagnamens">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// TODO: Actual tests; wildcards.
test(function() {
  assert_false(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof NodeList, "NodeList")
  assert_true(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof HTMLCollection, "HTMLCollection")
  var firstCollection = document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html"),
      secondCollection = document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html")
  assert_true(firstCollection !== secondCollection || firstCollection === secondCollection,
              "Caching is allowed.")
})
</script>
